First real-world usable prototype!
Reads config in from /sdcard/gabeldorsche.config (see
the sample gabeldorsche.config for how to use)

Notification vibe determined on the basis of the originating
package.

If the "sender" of the notification (not the package, but the entity
behind it, e.g. a person) matches a configured sender, the
characteristic vibe for that sender happens shortly after the
notification vibe.

If the text of the notification matches the urgent regex (suggested
hot-term: #urgent), the urgent vibe is sent before anything else.